# Admin-a-main/Admin-a-main/lib/api-spec/openapi.yaml
openapi: "3.0.3"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /rides:
    get:
      operationId: listRides
      tags: [rides]
      summary: List all rides
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Ride list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideListResponse"
  /rides/{id}:
    get:
      operationId: getRide
      tags: [rides]
      summary: Get ride by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ride details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideDetail"
        "404":
          description: Ride not found
    patch:
      operationId: updateRide
      tags: [rides]
      summary: Update ride (admin override)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideUpdate"
      responses:
        "200":
          description: Updated ride
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
  /rides/{id}/cancel:
    post:
      operationId: cancelRide
      tags: [rides]
      summary: Force cancel a ride
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelRideInput"
      responses:
        "200":
          description: Ride cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, phone, status, isVerified, totalRides, totalSpending, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        profileImage:
          type: string
        status:
          type: string
          enum: [active, inactive, banned, suspended]
        isVerified:
          type: boolean
        isBlockedFromBooking:
          type: boolean
        totalRides:
          type: integer
        cancelledRides:
          type: integer
        totalSpending:
          type: number
        walletBalance:
          type: number
        fcmToken:
          type: string
        lastActiveAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Driver:
      type: object
      required: [id, name, phone, status, approvalStatus, isOnline, vehicleType, totalRides, totalEarnings, rating, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        profileImage:
          type: string
        status:
          type: string
          enum: [active, pending, suspended, banned]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        isOnline:
          type: boolean
        vehicleType:
          type: string
          enum: [auto, bike, cab]
        vehicleNumber:
          type: string
        vehicleModel:
          type: string
        totalRides:
          type: integer
        totalEarnings:
          type: number
        rating:
          type: number
        walletBalance:
          type: number
        fcmToken:
          type: string
        lastActiveAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Ride:
      type: object
      required: [id, status, fare, paymentMethod, paymentStatus, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        driverId:
          type: string
        userName:
          type: string
        userPhone:
          type: string
        driverName:
          type: string
        driverPhone:
          type: string
        pickupAddress:
          type: string
        dropAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        dropLat:
          type: number
        dropLng:
          type: number
        distanceKm:
          type: number
        fare:
          type: number
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        paymentMethod:
          type: string
          enum: [cash, wallet, online]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        cancelReason:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    RideDetail:
      allOf:
        - $ref: "#/components/schemas/Ride"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            driver:
              $ref: "#/components/schemas/Driver"
    RideUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        fare:
          type: number
        driverId:
          type: string
    CancelRideInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
        refund:
          type: boolean
...
    RideListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ride"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...